Add simple authentication loop using continue and break
This program repeatedly asks for a name and password using a while loop. It only allows access if the name is 'Joe' and the password is 'swordfish'. Demonstrates the use of continue to skip to the next iteration and break to exit the loop upon successful authentication. Based on an example from "Automate the Boring Stuff with Python."